@@ -1,6 +1,6 @@
 {
   "name": "@binance-chain/javascript-sdk",
-  "version": "3.0.2",
+  "version": "3.0.2-alpha.8",
   "license": "Apache-2.0",
   "main": "lib/index.js",
   "typings": "typings/index.d.ts",
@@ -12,6 +12,7 @@
     "build:docs": "rimraf docs && typedoc --plugin typedoc-plugin-markdown --excludeExternals --excludeNotDocumented --readme none",
     "prepublishOnly": "npm run build",
     "clean": "rimraf lib typings",
+    "lint": "eslint '**/*.{ts,js}'",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,css,json,md,html,yml}\""
   },
   "dependencies": {
@@ -75,6 +76,8 @@
     "@types/pumpify": "^1.4.1",
     "@types/tiny-secp256k1": "^1.0.0",
     "@types/uuid": "^7.0.0",
+    "@typescript-eslint/eslint-plugin": "^3.0.0",
+    "@typescript-eslint/parser": "^3.1.0",
     "babel-core": "^7.0.0-0",
     "babel-eslint": "^10.0.2",
     "babel-jest": "^23.6.0",
@@ -87,17 +90,20 @@
     "bufferutil": "^4.0.1",
     "builtin-modules": "^3.1.0",
     "cross-env": "^5.2.0",
-    "eslint": "^5.11.1",
+    "eslint": "^7.1.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-prettier": "^3.1.3",
     "http-server": "^0.11.1",
-    "husky": "^3.0.0",
+    "husky": "^4.2.5",
     "jest": "^25.1.0",
     "prettier": "^2.0.4",
     "qunit": "^2.9.2",
     "rimraf": "^3.0.2",
     "ts-jest": "^25.2.1",
-    "typedoc": "^0.17.4",
+    "typedoc": "^0.17.7",
     "typedoc-plugin-markdown": "^2.2.17",
-    "typescript": "^3.8.3",
+    "typescript": "^3.9.3",
     "utf-8-validate": "^5.0.2"
   },
   "files": [
@@ -108,5 +114,10 @@
     "**/**/handlebars": "^4.1.2",
     "**/**/ecstatic": "^3.3.2",
     "**/**/mem": "^4.0.0"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "yarn lint"
+    }
   }
 }
