@@ -1,13 +1,11 @@
-/**
- * @module amino.encode
- */
-
-import { string as VarString } from "protocol-buffers-encodings"
 import is from "is_js"
-import { UVarInt } from "./varint"
+import { string as VarString } from "protocol-buffers-encodings"
+
 import typeToTyp3 from "../../utils/encoderHelper"
 
-const sortObject = (obj: any): object | null => {
+import { UVarInt } from "./varint"
+
+const sortObject = (obj: any): any => {
   if (obj === null) return null
   if (typeof obj !== "object") return obj
   // arrays have typeof "object" in js!
@@ -22,12 +20,14 @@ const sortObject = (obj: any): object | null => {
 
 /**
  * encode number
+ * @category amino
  * @param num
  */
 export const encodeNumber = (num: number) => UVarInt.encode(num)
 
 /**
  * encode bool
+ * @category amino
  * @param b
  */
 export const encodeBool = (b: boolean) =>
@@ -35,15 +35,17 @@ export const encodeBool = (b: boolean) =>
 
 /**
  * encode string
+ * @category amino
  * @param str
  */
 export const encodeString = (str: string) => {
-  var buf = Buffer.alloc(VarString.encodingLength(str))
+  const buf = Buffer.alloc(VarString.encodingLength(str))
   return VarString.encode(str, buf, 0)
 }
 
 /**
  * encode time
+ * @category amino
  * @param value
  */
 export const encodeTime = (value: string | Date) => {
@@ -65,6 +67,7 @@ export const encodeTime = (value: string | Date) => {
 }
 
 /**
+ * @category amino
  * @param obj -- {object}
  * @return bytes {Buffer}
  */
@@ -73,6 +76,7 @@ export const convertObjectToSignBytes = (obj: any) =>
 
 /**
  * js amino MarshalBinary
+ * @category amino
  * @param {Object} obj
  *  */
 export const marshalBinary = (obj: any) => {
@@ -83,6 +87,7 @@ export const marshalBinary = (obj: any) => {
 
 /**
  * js amino MarshalBinaryBare
+ * @category amino
  * @param {Object} obj
  *  */
 export const marshalBinaryBare = (obj: any) => {
@@ -93,6 +98,7 @@ export const marshalBinaryBare = (obj: any) => {
 
 /**
  * This is the main entrypoint for encoding all types in binary form.
+ * @category amino
  * @param {*} js data type (not null, not undefined)
  * @param {Number} field index of object
  * @param {Boolean} isByteLenPrefix
@@ -137,6 +143,7 @@ export const encodeBinary = (
 
 /**
  * prefixed with bytes length
+ * @category amino
  * @param {Buffer} bytes
  * @return {Buffer} with bytes length prefixed
  */
@@ -146,7 +153,7 @@ export const encodeBinaryByteArray = (bytes: Buffer) => {
 }
 
 /**
- *
+ * @category amino
  * @param {Object} obj
  * @return {Buffer} with bytes length prefixed
  */
@@ -184,6 +191,7 @@ export const encodeObjectBinary = (obj: any, isByteLenPrefix?: boolean) => {
 }
 
 /**
+ * @category amino
  * @param {Number} fieldNum object field index
  * @param {Array} arr
  * @param {Boolean} isByteLenPrefix
