@@ -1,6 +1,6 @@
-import { BaseMsg, Msg, SignMsg } from "../base"
 import * as crypto from "../../../crypto"
 import { AminoPrefix } from "../../tx"
+import { BaseMsg, Msg, SignMsg } from "../base"
 
 export interface SignedCancelOrder extends SignMsg {
   sender: string
@@ -31,7 +31,7 @@ export class CancelOrderMsg extends BaseMsg {
     const signMsg: SignedCancelOrder = {
       sender: this.address,
       symbol: this.symbol,
-      refid: this.orderId
+      refid: this.orderId,
     }
 
     return signMsg
@@ -42,7 +42,7 @@ export class CancelOrderMsg extends BaseMsg {
       sender: crypto.decodeAddress(this.address),
       symbol: this.symbol,
       refid: this.orderId,
-      aminoPrefix: AminoPrefix.CancelOrderMsg
+      aminoPrefix: AminoPrefix.CancelOrderMsg,
     }
 
     return data
@@ -53,7 +53,7 @@ export class CancelOrderMsg extends BaseMsg {
       sender: Buffer.from(""),
       symbol: "",
       refid: "",
-      aminoPrefix: AminoPrefix.CancelOrderMsg
+      aminoPrefix: AminoPrefix.CancelOrderMsg,
     }
   }
 }
