@@ -1,7 +1,9 @@
-import { BaseMsg, Msg, SignMsg } from "./"
+import Big, { BigSource } from "big.js"
+
 import * as crypto from "../../crypto"
 import { AminoPrefix } from "../tx"
-import Big, { BigSource } from "big.js"
+
+import { BaseMsg, Msg, SignMsg } from "./"
 
 export interface Coin {
   denom: string
@@ -30,6 +32,7 @@ export interface SendData extends Msg {
 }
 
 /**
+ * @ignore
  * Only support transfers of one-to-one, one-to-many
  */
 export class SendMsg extends BaseMsg {
