@@ -31,7 +31,7 @@ export class IssueTokenMsg extends BaseMsg {
   getSignMsg() {
     const signMsg: SignedIssueTokenMsg = {
       from: this.from,
-      ...this.params
+      ...this.params,
     }
 
     return signMsg
@@ -44,7 +44,7 @@ export class IssueTokenMsg extends BaseMsg {
       symbol: this.params.symbol,
       total_supply: this.params.total_supply,
       mintable: this.params.mintable,
-      aminoPrefix: AminoPrefix.IssueMsg
+      aminoPrefix: AminoPrefix.IssueMsg,
     }
 
     return data
@@ -57,7 +57,7 @@ export class IssueTokenMsg extends BaseMsg {
       symbol: "",
       total_supply: 0,
       mintable: false,
-      aminoPrefix: AminoPrefix.IssueMsg
+      aminoPrefix: AminoPrefix.IssueMsg,
     }
   }
 }
