@@ -1,7 +1,8 @@
+import Big from "big.js"
+
 import { BaseMsg, Msg, SignMsg } from "../"
 import * as crypto from "../../../crypto"
 import { AminoPrefix } from "../../tx"
-import Big from "big.js"
 
 export interface SignedMintTokenMsg extends SignMsg {
   from: string
@@ -24,7 +25,7 @@ export class MintTokenMsg extends BaseMsg {
   constructor({
     address,
     sybmol,
-    amount
+    amount,
   }: {
     address: string
     sybmol: string
@@ -40,7 +41,7 @@ export class MintTokenMsg extends BaseMsg {
     const signMsg: SignedMintTokenMsg = {
       from: this.from,
       amount: Number(new Big(this.amount).mul(Math.pow(10, 8)).toString()),
-      symbol: this.symbol
+      symbol: this.symbol,
     }
 
     return signMsg
@@ -51,7 +52,7 @@ export class MintTokenMsg extends BaseMsg {
       from: crypto.decodeAddress(this.from),
       symbol: this.symbol,
       amount: Number(new Big(this.amount).mul(Math.pow(10, 8)).toString()),
-      aminoPrefix: AminoPrefix.MintMsg
+      aminoPrefix: AminoPrefix.MintMsg,
     }
 
     return data
@@ -62,7 +63,7 @@ export class MintTokenMsg extends BaseMsg {
       from: Buffer.from(""),
       symbol: "",
       amount: 0,
-      aminoPrefix: AminoPrefix.MintMsg
+      aminoPrefix: AminoPrefix.MintMsg,
     }
   }
 }
