@@ -1,7 +1,6 @@
-import { BaseMsg, Msg, SignMsg } from "../"
+import { BaseMsg, Msg, SignMsg, Coin } from "../"
 import * as crypto from "../../../crypto"
 import { AminoPrefix } from "../../tx"
-import { Coin } from "../"
 
 export interface SignedTimeReLockMsg extends SignMsg {
   from: string
@@ -32,7 +31,7 @@ export class TimeReLockMsg extends BaseMsg {
     time_lock_id,
     description,
     amount,
-    lock_time
+    lock_time,
   }: {
     address: string
     time_lock_id: number
@@ -54,7 +53,7 @@ export class TimeReLockMsg extends BaseMsg {
       time_lock_id: this.time_lock_id,
       amount: this.amount,
       description: this.description,
-      lock_time: this.lock_time
+      lock_time: this.lock_time,
     }
 
     return signMsg
@@ -67,7 +66,7 @@ export class TimeReLockMsg extends BaseMsg {
       description: this.description,
       amount: this.amount,
       lock_time: this.lock_time,
-      aminoPrefix: AminoPrefix.TimeRelockMsg
+      aminoPrefix: AminoPrefix.TimeRelockMsg,
     }
 
     return data
@@ -79,7 +78,7 @@ export class TimeReLockMsg extends BaseMsg {
       description: "",
       amount: 0,
       lock_time: 0,
-      aminoPrefix: AminoPrefix.TimeRelockMsg
+      aminoPrefix: AminoPrefix.TimeRelockMsg,
     }
   }
 }
