@@ -19,7 +19,7 @@ export class TimeUnlockMsg extends BaseMsg {
 
   constructor({
     address,
-    time_lock_id
+    time_lock_id,
   }: {
     address: string
     time_lock_id: number
@@ -32,7 +32,7 @@ export class TimeUnlockMsg extends BaseMsg {
   getSignMsg() {
     const signMsg: SignedTimeUnlockMsg = {
       from: this.from,
-      time_lock_id: this.time_lock_id
+      time_lock_id: this.time_lock_id,
     }
 
     return signMsg
@@ -42,7 +42,7 @@ export class TimeUnlockMsg extends BaseMsg {
     const data: TimeUnlockData = {
       from: crypto.decodeAddress(this.from),
       time_lock_id: this.time_lock_id,
-      aminoPrefix: AminoPrefix.TimeUnlockMsg
+      aminoPrefix: AminoPrefix.TimeUnlockMsg,
     }
 
     return data
@@ -52,7 +52,7 @@ export class TimeUnlockMsg extends BaseMsg {
     return {
       from: Buffer.from(""),
       time_lock_id: 0,
-      aminoPrefix: AminoPrefix.TimeUnlockMsg
+      aminoPrefix: AminoPrefix.TimeUnlockMsg,
     }
   }
 }
